+++ dbt_platform_helper/commands/conduit.py
@@ -1,13 +1,13 @@
 import click
 
 from dbt_platform_helper.domain.conduit import Conduit
-from dbt_platform_helper.providers.aws import SecretNotFoundError
 from dbt_platform_helper.providers.copilot import CONDUIT_ADDON_TYPES
 from dbt_platform_helper.providers.copilot import AddonNotFoundError
 from dbt_platform_helper.providers.copilot import CreateTaskTimeoutError
 from dbt_platform_helper.providers.copilot import InvalidAddonTypeError
 from dbt_platform_helper.providers.copilot import NoClusterError
 from dbt_platform_helper.providers.copilot import ParameterNotFoundError
+from dbt_platform_helper.providers.secrets import SecretNotFoundError
 from dbt_platform_helper.utils.application import load_application
 from dbt_platform_helper.utils.click import ClickDocOptCommand
 from dbt_platform_helper.utils.versioning import (
