+++ dbt_platform_helper/providers/secrets.py
@@ -1,17 +1,9 @@
 import json
 import urllib
 
-
-# TODO exceptions
-class AWSError(Exception):
-    pass
+from dbt_platform_helper.exceptions import SecretNotFoundError
 
 
-class SecretNotFoundError(AWSError):
-    pass
-
-
-# TODO extract some business knowledge
 def get_postgres_connection_data_updated_with_master_secret(
     ssm_client, secrets_manager_client, parameter_name, secret_arn
 ):
