+++ tests/platform_helper/test_command_conduit.py
@@ -5,12 +5,12 @@ import pytest
 from click.testing import CliRunner
 
 from dbt_platform_helper.commands.conduit import conduit
-from dbt_platform_helper.providers.aws import SecretNotFoundError
 from dbt_platform_helper.providers.copilot import AddonNotFoundError
 from dbt_platform_helper.providers.copilot import CreateTaskTimeoutError
 from dbt_platform_helper.providers.copilot import InvalidAddonTypeError
 from dbt_platform_helper.providers.copilot import NoClusterError
 from dbt_platform_helper.providers.copilot import ParameterNotFoundError
+from dbt_platform_helper.providers.secrets import SecretNotFoundError
 
 
 @pytest.mark.parametrize(
